refactor: cleanup unused components, add visual toggle to parts on select
