+++ apps/www/components/component-explorer.tsx
@@ -1,89 +1,41 @@
-"use client"
+import { Box, Heading, Text } from "@chakra-ui/react"
+import { ComponentExplorerSidebar } from "./component-explorer-sidebar"
+import { ExamplePreview } from "./example"
 
-import { Box, Flex, Heading, Text } from "@chakra-ui/react"
-import * as anatomies from "@chakra-ui/react/anatomy"
-import { useCallback } from "react"
-import { kebabCase } from "scule"
-
-const normalizeComponentName = (name: string) => {
-  return name.split("-")[0].toLowerCase()
-}
-
-export const ComponentExplorerSidebar = ({
-  componentName,
+export function ComponentExplorer({
+  name,
+  className,
 }: {
-  componentName: string
-}) => {
-  const normalizedName = normalizeComponentName(componentName)
-  const anatomy =
-    anatomies[`${normalizedName}Anatomy` as keyof typeof anatomies]
-
-  const highlightPart = useCallback(
-    (part: string, action: "add" | "remove") => {
-      const kebabPart = kebabCase(part)
-      const selector = `[data-scope=${normalizedName}][data-part=${kebabPart}]`
-      const preview = document.getElementById("component-preview")
-      if (!preview) return
-
-      const elements = preview.querySelectorAll(
-        selector,
-      ) as NodeListOf<HTMLElement>
-
-      for (const el of elements) {
-        if (action === "add") {
-          el.style.outline = "2px solid red"
-          el.style.outlineOffset = "2px"
-        } else {
-          el.style.outline = ""
-          el.style.outlineOffset = ""
-        }
-      }
-    },
-    [normalizedName],
-  )
-
-  if (!anatomy) {
-    return (
-      <Box p={5}>
-        <Text>No anatomy found for {componentName}</Text>
-      </Box>
-    )
-  }
-
+  name: string
+  className?: string
+}) {
   return (
-    <Box p={5} borderLeft="1px solid" borderColor="border" bg="bg">
-      <Heading size="sm" mb={2} color="fg">
-        Component Anatomy
-      </Heading>
-      <Text fontSize="sm" mb={4} color="fg.muted">
-        Hover or click to highlight parts
-      </Text>
-
-      <Flex wrap="wrap" gap={2}>
-        {anatomy.keys().map((key) => (
-          <Box
-            key={key}
-            px={3}
-            py={2}
-            borderRadius="md"
-            cursor="pointer"
-            bg="bg.subtle"
-            border="1px solid"
-            borderColor="border.subtle"
-            _hover={{
-              bg: "bg.emphasized",
-              borderColor: "border.emphasized",
-            }}
-            onMouseEnter={() => highlightPart(key, "add")}
-            onMouseLeave={() => highlightPart(key, "remove")}
-            onClick={() => highlightPart(key, "add")}
-          >
-            <Text fontSize="sm" fontWeight="medium" textTransform="capitalize">
-              {key}
-            </Text>
-          </Box>
-        ))}
-      </Flex>
+    <Box
+      display="grid"
+      gridTemplateColumns={{ base: "1fr", lg: "1fr 260px" }}
+      minH="480px"
+      className={className}
+    >
+      <Box p={5}>
+        <Heading size="sm" mb={3} color="fg">
+          Preview
+        </Heading>
+        <Box
+          id="component-preview"
+          p={6}
+          minH="xs"
+          display="flex"
+          alignItems="center"
+          justifyContent="center"
+          bg="bg.subtle"
+          borderRadius="sm"
+          border="1px solid"
+          borderColor="border.subtle"
+        >
+          <ExamplePreview name={name} />
+        </Box>
+      </Box>
+      <ComponentExplorerSidebar componentName={name} />
     </Box>
   )
 }
