+++ apps/www/components/mdx-content.tsx
@@ -15,7 +15,7 @@ import { SpacingTokenDoc } from "compositions/lib/spacing-token-doc"
 import * as TypographyDocs from "compositions/lib/typography-token-doc"
 import { ZIndexTokenDoc } from "compositions/lib/z-index-token-doc"
 import * as runtime from "react/jsx-runtime"
-import { ComponentExplorerWrapper } from "./component-explorer-wrapper"
+import { ComponentExplorer } from "./component-explorer"
 import {
   Example,
   ExampleCode,
@@ -136,7 +136,7 @@ const sharedComponents = {
     )
   },
   Explorer(props: { name: string }) {
-    return <ComponentExplorerWrapper name={props.name} />
+    return <ComponentExplorer name={props.name} />
   },
 }
